Invoice_App/requirements.txt
fastapi
uvicorn
transformers
torch
pillow
bs4
langchain
langchain-huggingface
